+++ rust/agama-lib/src/install_settings.rs
@@ -1,6 +1,7 @@
 //! Configuration settings handling
 //!
 //! This module implements the mechanisms to load and store the installation settings.
+use crate::network::NetworkSettings;
 use crate::settings::{SettingObject, SettingValue, Settings};
 use agama_derive::Settings;
 use serde::{Deserialize, Serialize};
@@ -241,32 +242,6 @@ pub struct RootUserSettings {
     pub ssh_public_key: Option<String>,
 }
 
-/// Network settings for installation
-#[derive(Debug, Default, Settings, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct NetworkSettings {
-    /// Connections to use in the installation
-    #[collection_setting]
-    pub connections: Vec<NetworkConnection>,
-}
-
-#[derive(Debug, Default, Serialize, Deserialize)]
-pub struct NetworkConnection {
-    pub id: String,
-}
-
-impl TryFrom<SettingObject> for NetworkConnection {
-    type Error = &'static str;
-
-    fn try_from(value: SettingObject) -> Result<Self, Self::Error> {
-        match value.0.get("id") {
-            Some(name) => Ok(NetworkConnection {
-                id: name.clone().try_into()?,
-            }),
-            None => Err("'id' key not found"),
-        }
-    }
-}
 /// Storage settings for installation
 #[derive(Debug, Default, Settings, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
